PROXY-8: Refactor code Divide the application into layers - add the Delivery layer - add route to the Delivery layer - add the UseCases layer - add a Repository layer
