+++ package.json
@@ -2,11 +2,11 @@
   "name": "proxy",
   "version": "1.0.0",
   "description": "Intermediate server that acts as an intermediary between the user and the target server",
-  "main": "index.js",
+  "main": "src/server.js",
   "type": "module",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "start_server": "node --watch server.js"
+    "start_server": "node --watch src/server.js"
   },
   "author": "a.asenchik",
   "license": "ISC",
